# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/auth/google/callback": {
      "get": {
        "operationId": "AuthController_googleCallback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/google/link-confirm": {
      "post": {
        "operationId": "AuthController_confirmLink",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/phone/send-otp": {
      "post": {
        "operationId": "AuthController_sendPhoneOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/phone/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyPhoneOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/email/send-otp": {
      "post": {
        "operationId": "AuthController_sendEmailOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/email/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyEmailOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/username/check": {
      "get": {
        "operationId": "AuthController_checkUsername",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/google/complete-profile": {
      "post": {
        "operationId": "AuthController_completeProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Certifine API",
    "description": "C2C marketplace with per-category third-party authentication. Generated from the running app — do not hand-edit, and do not hand-edit the client generated from it.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "emailOtp": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "email",
          "displayName",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "password"
        ]
      },
      "ConfirmLinkDto": {
        "type": "object",
        "properties": {
          "linkToken": {
            "type": "string"
          }
        },
        "required": [
          "linkToken"
        ]
      },
      "SendOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "REGISTER_PHONE",
              "CHANGE_PHONE"
            ]
          }
        },
        "required": [
          "phone",
          "purpose"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "REGISTER_PHONE",
              "CHANGE_PHONE"
            ]
          }
        },
        "required": [
          "phone",
          "code",
          "purpose"
        ]
      },
      "SendEmailOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "REGISTER_EMAIL",
              "VERIFY_EMAIL"
            ]
          }
        },
        "required": [
          "email",
          "purpose"
        ]
      },
      "VerifyEmailOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "code",
          "purpose"
        ]
      },
      "CompleteProfileDto": {
        "type": "object",
        "properties": {
          "completeToken": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "useSuggestedAvatar": {
            "type": "boolean"
          }
        },
        "required": [
          "completeToken",
          "displayName",
          "useSuggestedAvatar"
        ]
      },
...
    }
  }
}